more err handling

+++ torles.php
@@ -1,5 +1,11 @@
 <?php
 require("kapcsolat.php");
+if (!isset($_GET["id"]))
+{
+    echo "Id paraméter hiányzik";
+    die();
+}
+
 $sql = "DELETE FROM dolgozok WHERE id like {$_GET['id']}";
 if($dbconn->query($sql))
 {
